+++ src/main/java/com/kdev5/cleanpick/contract/service/ContractService.java
@@ -1,9 +1,9 @@
 package com.kdev5.cleanpick.contract.service;
 
 
-import com.kdev5.cleanpick.contract.dto.request.ContractRequestDto;
-import com.kdev5.cleanpick.contract.dto.response.OneContractResponseDto;
-import com.kdev5.cleanpick.contract.dto.response.RoutineContractResponseDto;
+import com.kdev5.cleanpick.contract.service.dto.request.ContractRequestDto;
+import com.kdev5.cleanpick.contract.service.dto.response.OneContractResponseDto;
+import com.kdev5.cleanpick.contract.service.dto.response.RoutineContractResponseDto;
 import jakarta.validation.Valid;
 
 public interface ContractService {
