+++ src/main/java/com/kdev5/cleanpick/contract/service/ContractServiceImpl.java
@@ -12,9 +12,9 @@ import com.kdev5.cleanpick.contract.domain.Contract;
 import com.kdev5.cleanpick.contract.domain.ContractDetail;
 import com.kdev5.cleanpick.contract.domain.RoutineContract;
 import com.kdev5.cleanpick.contract.domain.enumeration.ContractStatus;
-import com.kdev5.cleanpick.contract.dto.request.ContractRequestDto;
-import com.kdev5.cleanpick.contract.dto.response.OneContractResponseDto;
-import com.kdev5.cleanpick.contract.dto.response.RoutineContractResponseDto;
+import com.kdev5.cleanpick.contract.service.dto.request.ContractRequestDto;
+import com.kdev5.cleanpick.contract.service.dto.response.OneContractResponseDto;
+import com.kdev5.cleanpick.contract.service.dto.response.RoutineContractResponseDto;
 import com.kdev5.cleanpick.contract.domain.exception.ContractException;
 import com.kdev5.cleanpick.contract.infra.*;
 import com.kdev5.cleanpick.customer.domain.Customer;
@@ -30,7 +30,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.DayOfWeek;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
