+++ src/main/java/com/kdev5/cleanpick/contract/service/dto/response/RoutineContractResponseDto.java
@@ -1,4 +1,4 @@
-package com.kdev5.cleanpick.contract.dto.response;
+package com.kdev5.cleanpick.contract.service.dto.response;
 
 import com.kdev5.cleanpick.contract.domain.RoutineContract;
 import lombok.Getter;
